added mailboy my module that will send the email.
